Allow for asynchronous loading of audio
Added code and styles to allow for the application to start instead of
waiting for the external audio files to be loaded. This only applies
the vinyl sounds at this point.

Added a deactivated class for the buttons for when the sound hasn't
been loaded and also moved the loadSound call into the setup portion
of the code. Some other minor changes were made to verify that sound
has loaded before allowing dependent code to execute.